add read formats consistency test
Checks that TABLEDATA and BINARY2 tables with the same
content look the same when read by this package. BINARY file is added, but reading isn't implemented yet.

The test files are included in this commit, as is a shell script
that can generate them using the STILTS package
(http://www.starlink.ac.uk/stilts/)


Co-authored-by: Mark Taylor <m.b.taylor@bristol.ac.uk>